Refactor activity CLI to use processing helpers
