+++ library/cli/parser.py
@@ -16,7 +16,7 @@ from typing import Any
 from pydantic import ValidationError
 
 from .. import log
-from ..config import Config, ConfigError, load_config
+from ..utils.config import Config, ConfigError, load_config
 from ..logging_setup import Logger, LoggerConfig
 from ..logging_setup import configure_logger as _configure_logger
 from ..version import require_python_version
