+++ library/processing/__init__.py
@@ -1,0 +1,5 @@
+"""Processing helpers for ChEMBL data acquisition."""
+
+from . import activity  # noqa: F401
+
+__all__ = ["activity"]
