+++ library/utils/__init__.py
@@ -1,0 +1,5 @@
+"""Utility helpers for command line interfaces."""
+
+from . import config
+
+__all__ = ["config"]
