+++ library/utils/config.py
@@ -1,0 +1,13 @@
+"""Convenience re-exports for configuration helpers used by CLI scripts."""
+
+from __future__ import annotations
+
+from library.config import Config, ConfigError, ensure_dirs, load_config, print_config
+
+__all__ = [
+    "Config",
+    "ConfigError",
+    "ensure_dirs",
+    "load_config",
+    "print_config",
+]
